7/4 12:42

+++ C01/ex01/zombieHorde.cpp
@@ -24,9 +24,8 @@ Zombie* zombieHorde( int N, std::string name )
 	Zombie* horde = new Zombie[N];
 	std::string testname;
 
-	for (size_t i = 0; i < N; i++)
+	for (int i = 0; i < N; i++)
 		horde[i].SetNameZombie(name);
 		
-	
 	return (horde);
 }
